#3432 resolved comments on drag and drop feature
